Initial support for reaction redaction
With this patch, EventAnnotation events are now stored in gomuk's local
history.
Redacted reaction events are treated the same way as any other redacted
event in th history, but are ignored by the Message's event parser, in
order not to display them in the timeline.

There is still a problem with redacting reactions that come from the
initial sync, as it seems events are received in an aggregated form.

Also some of the code could be refactored.